Fixing release number(s) [modified changelog].
Perhaps it's bad practice to modify the changelog, but since the package
hasn't actually been released yet it seems like a reasonable thing to do
to make it correct. Also updated the date to today so that it is more
correct (the date that was there was well before the actual package date
that I requested review for).